combine 3 json util files into one jsonToCssScssTs
