--- design-tokens/utils/index.js
@@ -1,2 +1,0 @@
-export * from './jsonToScss';
-export * from './jsonToTs';
