--- design-tokens/utils/jsonUtils.js
@@ -1,50 +1,0 @@
-function isTrueObject(obj) {
-	return obj && typeof obj === 'object' && !Array.isArray(obj);
-}
-function isArray(obj) {
-	return obj && Array.isArray(obj);
-	// return obj && typeof obj === 'object' && Array.isArray(obj);
-}
-
-function hasChildObjects(obj) {
-	if (!isTrueObject(obj)) {
-		return false;
-	}
-	const childVals = Object.values(obj);
-	return childVals.some((kid) => typeof kid === 'object');
-}
-
-/* based off https://stackoverflow.com/a/2970667 */
-/**
- * @param {string} str
- * @param {object} options
- * @description convert string of space-separated or dash-separated words to camelCase
- * Returned string will start with lowercase letter unless option.titleCase is true.
- */
-function camelize(str, options) {
-	if (typeof str !== 'string') {
-		return '';
-	}
-	const { titleCase } = options;
-	return str
-		.replace(/-/g, ' ')
-		.replace(/(?:^\w|[A-Z]|\b\w|\s+)/g, function (match, index) {
-			if (+match === 0) return '';
-			if (titleCase) {
-				return match.toUpperCase();
-			}
-			return index === 0 ? match.toLowerCase() : match.toUpperCase();
-		});
-}
-
-function titleCase(str) {
-	return camelize(str, { titleCase: true });
-}
-
-module.exports = {
-	camelize,
-	hasChildObjects,
-	isArray,
-	isTrueObject,
-	titleCase,
-};
